+++ index.html
@@ -109,7 +109,9 @@
                 </li>
             </template>
     </div>
-    <script src="./scripts/script.js"></script>
+    <script type="module" src="./scripts/validation.js"></script>
+    <script  type="module"  src="./scripts/script.js"></script>
+
 </body>
 </html>
